Story 7.5: Graceful Handling of Orientation Changes

+++ app/src/main/java/com/nguyenmoclam/pexelssample/data/remote/model/PexelsPhotoDto.kt
@@ -18,7 +18,7 @@ data class PexelsPhotoDto(
     @SerialName("photographer_url")
     val photographerUrl: String,
     @SerialName("photographer_id")
-    val photographerId: Int,
+    val photographerId: Long,
     @SerialName("avg_color")
     val avgColor: String,
     @SerialName("src")
